stock_chart/requirements.txt
fastapi==0.116.0
tensorflow-cpu==2.19.0
opencv-python==4.12.0.88
uvicorn==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
pillow==11.3.0
numpy==2.1.3